### 2021-10-01 #3794 @whedon
[PRE REVIEW]: shar: A R package to analyze species-habitat associations using point pattern analysis
**Submitting author:** @mhesselbarth (<a href="http://orcid.org/0000-0003-1125-9918">Maximilian H.K. Hesselbarth</a>)
**Repository:** <a href="https://github.com/r-spatialecology/shar" target ="_blank">https://github.com/r-spatialecology/shar</a>
**Version:** v1.2.2
