### 2021-10-04 #3794 @Bisaloo
Hi @baddstats :wave:, this package appears to build upon spatstats (among other packages) to provide a more convenient/direct interface for researchers in ecology studying species-habitat associations. Would you be able to review the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03794/joss.03794/10.21105.joss.03794.pdf) and [R package](https://github.com/r-spatialecology/shar) for JOSS?
