### 2021-10-06 #3794 @Bisaloo
Hi @BenCretois :wave:, I saw your blog posts about spatial point processes in R, and after looking at your research interests, this seems like a very good fit for you. Would you be available to review the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03794/joss.03794/10.21105.joss.03794.pdf) and [R package](https://github.com/r-spatialecology/shar) for JOSS?
