### 2021-10-06 #3794 @BenCretois
@Bisaloo I would have loved to review the MS but I will have to defend my thesis in about 2 weeks so I may get a bit busy these days :-)
I can suggest Jorge: jorge.sicacha@ntnu.no who is very experienced in point processes!

Good luck for finding reviewers!
