### 2021-10-08 #3794 @Bisaloo
@mhesselbarth, this package is slightly outside my area of expertise and I don't have personal contacts with anyone working on these topic. Would you have suggestions of people potentially interested in reviewing this submission? Feel free to have a look at our [list of people that volunteered to review for JOSS](https://bit.ly/joss-reviewers) for inspiration.
