### 2021-10-08 #3794 @Bisaloo
Thanks for the quick answer and the suggestion!

:wave: @jeffreyevans, would you be available to review the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03794/joss.03794/10.21105.joss.03794.pdf) and [R package](https://github.com/r-spatialecology/shar) for JOSS?
