### 2021-10-13 #3811 @lionel68
Ok just made a quick first go through the manuscript and still have to check the functionalities of the package itself so right now I will focus on the text:

1. The species-habitat association derived from point pattern as used in `shar` are based on null models, there are however other ways to estimate these associations by directly modelling the intensity of the point pattern to habitat / covariates, I am thinking of [inlabru::lgcp](https://inlabru-org.github.io/inlabru/articles/web/2d_lgcp_covars.html) or [spatstat::kppm](https://www.rdocumentation.org/packages/spatstat/versions/1.64-1/topics/kppm). These other, more direct, approaches to estimate species-habitat association should be mentioned in the text
2.  The text is in some part a bit confusing and could be improved:
- Line 8: "Studying species-habitat associations is one tool to reveal ... ", I find it weird to refer to studies as tools, maybe replace tool with "approach" or similar?
- Line 15-19: I was a bit confused in this part about the ecological level, I usually envision point pattern analysis as being done within a species with **individuals** clustering in certain habitats and environmental conditions. Analysis of species-habitat associations is usually done within a species (as is the case in this package), I found these sentences therefore confusing concerning which levels where under interest here.
- Line 26-28: The definition of point pattern is rather terse, I would expand and add concrete examples to make it clearer to the readers which datasets would qualify as a point pattern and which not.

3. It would be interesting to have some ideas on which raster / point pattern size would give results in a reasonable amount of time with the package. Like how long does it takes to run 10'000 simulations of a raster with 1Mio pixels? And are there options / functionalities built in the package to run these computations on large dataset efficiently (parallel computing, GPU ...)

Voila that's it for now, I'll look at the package R functionalities in the next days.
