### 2021-10-15 5ec4058
:sparkles: openjournals/joss-reviews#3811 Adding list_to_randomized to allow parallelization
