### 2021-10-17 r-spatialecology/shar#75 @tretherington
Clear statement of dependancies
Regarding @openjournals/joss-reviews#3811 JOSS requires a "clearly-stated list of dependencies" at the moment you state in the README that "shar is mainly based on the spatstat (Baddeley et al. 2015) and raster (Hijmans 2017) package" but I think you should be specific as shar also requires the rcpp package.

I would suggest listing these somewhere, perhaps within the installation section.

I have spotted that you have an [open issue](https://github.com/r-spatialecology/shar/issues/71) about whether to include the rcpp requirement.  This is very much a style thing, but I prefer to have absolutely minimal package dependencies to make maintenance easier.  So you could help resolve this issue by removing that rcpp dependency - but a clear list of the remaining dependencies would still be helpful I think :smile:
