### 2021-10-17 r-spatialecology/shar#76 @tretherington
Community guidelines
Regarding @openjournals/joss-reviews#3811 JOSS requires "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" so I think you need to add something to the README to explain that.

I think I also forgot to include this for my JOSS paper :flushed:, so if it is helpful you could look at my [virtualNicheR](https://github.com/manaakiwhenua/virtualNicheR) package to see how I included that information.
