### 2021-10-17 r-spatialecology/shar#77 @tretherington
Include links to citations in README
Regarding @openjournals/joss-reviews#3811 I would have found it helpful to have links to the citations in the README.  Any chance you could please add the missing links you have in the paper to the README as well?  I think people who find your package via GitHub rather than via JOSS would find that helpful.
