### 2021-10-17 r-spatialecology/shar#78 @tretherington
Paper typos
Regarding @openjournals/joss-reviews#3811 I think there a couple of typos that you could correct:

Line 41: environmental conditions need to be broken by randomizing the data as **a** null model
Line 45: The first approach to simulate null model data, is to **randomize** the environmental data
