### 2021-10-17 r-spatialecology/shar#79 @tretherington
Test data examples potentially confusing?
Regarding @openjournals/joss-reviews#3811 I'm wondering if you could improve the test data examples slightly to avoid any confusion.

Ideally I think I should be able to run test data and get exactly the same results to confirm everything is working properly.  This is a bit challenging for you as the software uses data randomisation so you are unlikely to get **exactly** the same results.  However, I think it would be helpful if the conclusions of the results stays the same.

Everything seems to be fine with

``` r
results_habitat_association(pattern = species_a, raster = torus_trans)
## > Input: randomized raster | Quantile thresholds: negative < 0.025 - positive > 0.975
##   habitat count lo hi significance
## 1       1     9  2 14         n.s.
## 2       2    25  9 24     positive
## 3       3    27 11 26     positive
## 4       4     0 11 29     negative
## 5       5    12  5 18         n.s.
```

for which I get identical output in R.

But for

```r
results_habitat_association(pattern = reconstruction, raster = landscape_classified)
## > Input: randomized point pattern | Quantile thresholds: negative < 0.025 - positive > 0.975
##   habitat count    lo    hi significance
## 1       1     8  1.00 23.35         n.s.
## 2       2    22 25.85 49.10     negative
## 3       3    33 40.80 68.05     negative
## 4       4    19 44.00 71.10     negative
## 5       5   118 20.85 60.05     positive
```

I get the following result

```r
results_habitat_association(pattern = reconstruction, raster = landscape_classified)
> Input: randomized point pattern | Quantile thresholds: negative < 0.025 - positive > 0.975
  habitat count    lo    hi significance
1       1     8 27.90 48.10     negative
2       2    22 52.90 77.05     negative
3       3    33 29.00 57.30         n.s.
4       4    19 24.90 46.10     negative
5       5   118 11.85 26.00     positive
```

that has quite differrent lo and hi values, and different significance results for habitat classes 1 and 3.

I am assuming that all is running correctly, but I feel a bit nervous that my conclusion is different to that of the test data example.

I would suggest you:

1. Double-check that the output in the README is correct and up to date - I suppose there is a chance you changes the test data but forgot to update the README output.
2. Consider setting the random seed `set.seed()` as part of the example code so that the results will be the same.
3. Perhaps manipulate the test data such that the while the exact numbers may vary, at least the significance of the results is consistent.
