### 2021-10-17 r-spatialecology/shar#80 @tretherington
Why a point pattern approach to habitat analyses
Regarding @openjournals/joss-reviews#3811 requirement for a statement of need, I am not aware of any point pattern software in R that allows someone to do habitat analyses, so in that regard I am happy that this software fills a gap.

However, I think it would be helpful for me to understand if there are situations where this approach is beneficial over the multitude of other ways I could take xy point data and a raster and produce some sort of habitat/species distribution/ecological niche model.

I don't feel like I can articulate this question very well, so I do apologise for that :grimacing:, and I haven't had time to fully read and digest the cited papers, but I think it could be very helpful to potential users to understand any specific instances where these kinds of approaches are beneficial when compared to other similar presence-background approaches available in R packages such as dismo, biomod2, zoon, adehabitat, and the like.

I guess I would say that at the moment I don't see what the benefit of treating my xy location data as a point pattern and analysing it though shar.

I am mentioning this as if you can highlight that benefit somehow, then I think you will increase the likelihood of ecologists engaging with shar.
