+++ DESCRIPTION
@@ -23,19 +23,19 @@ Imports:
     grDevices,
     methods,
     raster,
+    Rcpp,
     spatstat.core,
     spatstat.geom,
     stats,
     utils,
-    Rcpp
 RoxygenNote: 7.1.2
 Suggests:
     covr,
     dplyr,
-    testthat (>= 2.1.0),
-    knitr,
+    future,
+    knitr, 
     rmarkdown,
-    spatstat
+    testthat (>= 2.1.0)
 VignetteBuilder: knitr
 Encoding: UTF-8
 LazyData: true
