+++ paper.bib
@@ -7,6 +7,14 @@
   address = {London}
 }
 
+@article{Bengtsson2021,
+  title = {A unifying framework for parallel and distributed processing in R using Futures},
+  author = {Bengtsson, H.},
+  year = {2021},
+  journal = {The R Journal},
+  doi = {10.32614/RJ-2021-048}
+}
+
 @article{Comita2007,
   title = {Developmental changes in habitat associations of tropical trees},
   author = {Comita, L.S. and Condit, R. and Hubbell, S.P.},
@@ -29,6 +37,31 @@
   doi = {10.1038/s41598-017-15789-5}
 }
 
+@article{Esser2015,
+  title = {Spatial scales of interactions among bacteria and between bacteria and the leaf surface},
+  author = {Esser, D.S. and Leveau, J.H.J. and Meyer, K.M. and Wiegand, K.},
+  year = {2015},
+  journal = {FEMS Microbiology Ecology},
+  volume = {91},
+  number = {3},
+  pages = {1--13},
+  doi = {10.1093/femsec/fiu034}
+}
+
+@article{Funwi-Gabga2012,
+  title = {Understanding the nesting spatial behaviour of gorillas in the Kagwene Sanctuary, Cameroon},
+  author = {{Funwi-Gabga}, N. and Mateu, J.},
+  year = {2012},
+  month = aug,
+  journal = {Stochastic Environmental Research and Risk Assessment},
+  volume = {26},
+  number = {6},
+  pages = {793--811},
+  issn = {1436-3240, 1436-3259},
+  doi = {10.1007/s00477-011-0541-1},
+  langid = {english}
+}
+
 @article{Furniss2017,
   title = {Reconciling niches and neutrality in a subalpine temperate forest},
   author = {Furniss, T.J. and Larson, A.J. and Lutz, J.A.},
@@ -49,6 +82,20 @@
   number = {2},
   pages = {349--362},
   doi = {10.1111/jvs.12090}
+}
+
+@article{Getzin2015,
+  title = {Adopting a spatially explicit perspective to study the mysterious fairy circles of Namibia},
+  author = {Getzin, S. and Wiegand, K. and Wiegand, T. and Yizhaq, H. and {von Hardenberg}, J. and Meron, E.},
+  year = {2015},
+  month = jan,
+  journal = {Ecography},
+  volume = {38},
+  number = {1},
+  pages = {1--11},
+  issn = {09067590},
+  doi = {10.1111/ecog.00911},
+  langid = {english}
 }
 
 @article{Guo2016,
@@ -90,6 +137,16 @@
   doi = {10.1073/pnas.0604666104}
 }
 
+@article{Klaas2000,
+  title = {The tempo and mode of gopher mound production in a tallgrass prairie remnant},
+  author = {Klaas, B.A. and Moloney, K.A. and Danielson, B.J.},
+  year = {2000},
+  journal = {Ecography},
+  volume = {23},
+  pages = {246--256},
+  doi = {10.1111/j.1600-0587.2000.tb00280.x}
+}
+
 @article{Kirkpatrick1983,
   title = {Optimization by simulated annealing},
   author = {Kirkpatrick, S. and Gelatt, C.D.Jr. and Vecchi, M.P.},
@@ -110,6 +167,14 @@
   number = {1},
   pages = {81--99},
   doi = {10.1006/jtbi.2000.2158}
+}
+
+@misc{RCoreTeam2021,
+  title = {R: A language and environment for statistical computing.},
+  author = {{R Core Team}},
+  year = {2021},
+  address = {Vienna (Austria)},
+  howpublished = {R Foundation for Statistical Computing}
 }
 
 @incollection{Tilman1993,
