### 2021-10-18 a3d2ae9
:fire: #75 #71 Remove Rcpp
