+++ R/reconstruct_pattern_cluster.R
@@ -168,7 +168,7 @@ reconstruct_pattern_cluster <- function(pattern,
     # difference between patterns
     difference <- simulated$n - pattern$n
     # id of points to remove
-    remove_points <- shar::rcpp_sample(x = seq_len(simulated$n), n = difference)
+    remove_points <- sample(x = seq_len(simulated$n), size = difference)
     # remove points
     simulated <- simulated[-remove_points]
 
@@ -238,8 +238,7 @@ reconstruct_pattern_cluster <- function(pattern,
     energy_df <- data.frame(i = seq(from = 1, to = max_runs, by = 1), energy = NA)
 
     # random ids of pattern
-    rp_id <- shar::rcpp_sample(x = seq_len(simulated_current$n), n = max_runs,
-                               replace = TRUE)
+    rp_id <- sample(x = seq_len(simulated_current$n), size = max_runs, replace = TRUE)
 
     # create random new points
     rp_coords <- spatstat.core::runifpoint(n = max_runs, nsim = 1, drop = TRUE,
