+++ man/shar.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/SHAR.R
+% Please edit documentation in R/shar-package.R
 \docType{package}
 \name{shar}
 \alias{shar}
