+++ DESCRIPTION
@@ -26,8 +26,7 @@ Imports:
     spatstat.core,
     spatstat.geom,
     stats,
-    utils,
-    Rcpp
+    utils
 RoxygenNote: 7.1.2
 Suggests:
     covr,
@@ -39,5 +38,3 @@ Suggests:
 VignetteBuilder: knitr
 Encoding: UTF-8
 LazyData: true
-LinkingTo: 
-    Rcpp
