### 2021-10-18 03c51e9
#77 Add reference links to README.Rmd
