+++ README.Rmd
@@ -56,7 +56,7 @@ library(raster)
 landscape_classified <- classify_habitats(raster = landscape, classes = 5)
 ```
 
-There are two possibilities to randomize the environmental data, both described in Harms et al. (2001). The first shifts the habitat map in all 4 cardinal directions around a torus. The second one assigns the habitat values to an empty map using a random walk algorithm. Both functions return a list with randomized rasters and the observed one. For more information on the methods, please click [here](https://r-spatialecology.github.io/shar/articles/background.html).
+There are two possibilities to randomize the environmental data, both described in Harms et al. (2001). The first shifts the habitat map in all 4 cardinal directions around a torus. The second one assigns the habitat values to an empty map using a random walk algorithm. Both functions return a list with randomized rasters and the observed one. For more information on the methods, please click [here](https://r-spatialecology.github.io/shar/articles/articles/background.html).
 
 ```{r habitat_random, eval = FALSE}
 torus_trans <- translate_raster(raster = landscape_classified, verbose = FALSE)
