### 2021-10-18 da30d0e
#76 Adding clearer statement about contributing
