+++ README.md
@@ -25,8 +25,8 @@ Positive or negative associations are present if the observed counts is
 higher or lower than the randomized counts (using quantile thresholds).
 Methods are mainly described in Plotkin et al. (2000), Harms et
 al. (2001) and Wiegand & Moloney (2014). **shar** is mainly based on
-the `spatstat` (Baddeley et al. 2015) and `raster` (Hijmans 2017)
-package.
+the [`spatstat`](http://spatstat.org) (Baddeley et al. 2015) and
+[`raster`](https://rspatial.org/raster/) (Hijmans 2017) package.
 
 ## Installation
 
@@ -45,11 +45,17 @@ And the development version from
 devtools::install_github("r-spatialecology/shar")
 ```
 
+This also automatically installs all non-base `R` package dependencies,
+namely the following packages: `classInt`, `raster`, `spatstat.core`,
+`spatstat.geom`.
+
 ## How to use shar
 
 ``` r
 library(shar)
 library(raster)
+
+set.seed(42)
 ```
 
 **shar** comes with build-in example data sets. `species_a` and
@@ -68,12 +74,12 @@ all 4 cardinal directions around a torus. The second one assigns the
 habitat values to an empty map using a random walk algorithm. Both
 functions return a list with randomized rasters and the observed one.
 For more information on the methods, please click
-[here](https://r-spatialecology.github.io/shar/articles/background.html).
+[here](https://r-spatialecology.github.io/shar/articles/articles/background.html).
 
 ``` r
-torus_trans <- translate_raster(raster = landscape_classified, verbose = FALSE)
+torus_trans <- translate_raster(raster = landscape_classified)
 
-random_walk <- randomize_raster(raster = landscape_classified, n_random = 39, verbose = FALSE)
+random_walk <- randomize_raster(raster = landscape_classified, n_random = 99)
 ```
 
 ``` r
@@ -85,13 +91,14 @@ plot_randomized_raster(torus_trans, n = 3, col = col)
 <img src="man/figures/README-plot_habitat-random-1.png" style="display: block; margin: auto;" />
 
 To randomize the point pattern, either use the Gamma test described by
-Plotkin et al. (2000) or pattern reconstruction (Tscheschel & Stoyan
-2006).
+Plotkin et al. (2000) or pattern reconstruction (Kirkpatrick et
+al. 1983; Tscheschel & Stoyan 2006).
 
 ``` r
-gamma_test <- fit_point_process(pattern = species_a, process = "cluster", n_random = 39, verbose = FALSE)
+gamma_test <- fit_point_process(pattern = species_b, process = "cluster", n_random = 99)
 
-reconstruction <- reconstruct_pattern_cluster(pattern = species_b, n_random = 39, verbose = FALSE) # takes some time
+# (this can takes some time)
+reconstruction <- reconstruct_pattern_cluster(pattern = species_b, n_random = 99, e_threshold = 0.05)
 ```
 
 Of course, there are several utility functions. For example, you can
@@ -107,21 +114,45 @@ plot_randomized_pattern(reconstruction, verbose = FALSE, ask = FALSE)
 ``` r
 calculate_energy(reconstruction, verbose = FALSE)
 ##  randomized_1  randomized_2  randomized_3  randomized_4  randomized_5 
-##    0.01716646    0.02080741    0.02469685    0.02252225    0.01905148 
+##    0.04766766    0.04981575    0.04924655    0.04881003    0.04994730 
 ##  randomized_6  randomized_7  randomized_8  randomized_9 randomized_10 
-##    0.02317256    0.01467501    0.02009006    0.01796639    0.01742050 
[…]
