+++ paper.md
@@ -29,24 +29,27 @@ The `shar` package builds on widely used `R` packages for spatial analyses and p
 
 # Statement of need
 
-Species-habitat associations are a result of certain species being specialized to certain abiotic conditions [@Tilman1993] and typically result in a clustering of individuals at suitable habitats [@Harms2001; @Comita2007].
+Species-habitat associations are a result of certain species being specialized to certain environmental conditions [@Tilman1993] and typically result in a clustering of individuals at suitable habitats [@Harms2001; @Comita2007].
 Thus, analyzing species-habitat associations can help to understand the importance of abiotic processes shaping the spatial distribution of species [@Garzon-Lopez2014].
-However, since biotic processes (e.g., competition, limited dispersal) can also lead to a clustering of species, analyses of species-habitat associations need to control for potential biotic processes because they result in a violation of the independence assumption of statistical tests such as the $\chi^2$ test [@Plotkin2000].
+However, since biotic processes (e.g., competition, limited dispersal) can also lead to a clustering of individuals, analyses of species-habitat associations need to control for potential biotic processes because they result in a violation of the independence assumption of similar statistical tests, such as the $\chi^2$ test, ordination methods, or canonical correspondence analysis [@Plotkin2000; @Harms2001].
+Simlar, also spatial autocorrelation of the environmental conditions could violate the independence assumption of the previously mentioned statistical tests [@Harms2001].
+Previous research has shown that violating the independence assumptions resulted in more (i.e., possible false-positive) species-habitat associations than the here presented, more conservative methods [@Plotkin2000; @Harms2001].
 
-Ecologists use the spatial distribution of ecological objects to infer the underlying processes that shaped their distribution.
+Ecologists use the spatial distribution of ecological objects to infer the underlying processes that shaped their distribution because spatial patterns can act as a "memory" of the processes that shaped them [@Velazquez2016].
 A spatial point pattern contains the fully mapped locations (in terms of *x*,*y* coordinates) of all individual objects in a normally two-dimensional study area and assumes that the object locations can be described by discrete points [@Wiegand2014; @Velazquez2016].
 For example, many studies use individual tree locations to infer the processes that determined their distribution [@Velazquez2016], and further examples include gopher mounds [@Klaas2000], gorilla nests [@Funwi-Gabga2012], "fairy circles" [@Getzin2015], or bacteria on leaves [@Esser2015].
 The `spatstat` package [@Baddeley2015] allows ecologists to access many methods of spatial point pattern analysis, such as summary functions and null model simulations.
 However, even though many ecological studies on species-habitat associations can be found in the literature [@John2007; @Garzon-Lopez2014; @Guo2016; @Yang2016; @Du2017; @Furniss2017], `spatstat` cannot be used to reveal species-habitat associations without larger programming efforts by the users.
 
 Thus, the `shar` package was developed to provide a tool set to analyze species-habitat associations of spatial point patterns.
+All methods in the `shar` package are designed for discrete environmental data and have the advantage of very few assumptions about the data characteristics.
 In order to make the package as accessible for as many people as possible, it builds on two of the most commonly used `R` packages for spatial data, namely the `spatstat` and `raster` packages [@Hijmans2019].
 
 # Methodological background
 
-To analyze species-habitat associations, potential dependencies between the object locations and the environmental conditions need to be broken by randomizing the data as a null model.
+To analyze species-habitat associations, potential interdependence between the object locations and the environmental conditions need to be broken by randomizing the data as a null model.
 Within the field of spatial point pattern analysis, there are two related approaches to break potential dependencies [@Plotkin2000; @Harms2001].
-Both require the spatial location of all objects, as well as raster data for environmental conditions.
[…]
