### 2021-10-18 r-spatialecology/shar#80 @mhesselbarth
5e7b241161f9a54c0c9cdac55b087e0f25a6ae46

I tried to add a few (short) explanation why the approach is interesting including:
- point pattern act as memory of processes that shaped them
- approach has very few assumptions of data characteristics because it automatically controls for spatial autocorrelation and dependence between points
- designed for discrete environmental data

I hope that answer your questions (at least to a degree...).
