### 2021-10-20 r-spatialecology/shar#81 @lionel68
Add precision in DESCRIPTION file
Regarding openjournals/joss-reviews#3811, shar deals with point pattern data linked to categorical (discrete) environment variables. This information on the data requirements should be in the DESCRIPTION file. Maybe:

`Therefore, information about the location of the species (as a point pattern) is needed together with environmental conditions (as a categorical raster).`
