### 2021-10-20 r-spatialecology/shar#82 @lionel68
Create plot methods for the shar object
Regarding issue openjournals/joss-reviews#3811, it would be easiest if plot methods were created for the objects derived from shar (i.e. rd_pat), so basically instead of having to call:

```
pattern_random <- fit_point_process(species_a, n_random = 19, process = "cluster")
plot_randomized_pattern(pattern_random)
```

One could simply do:

```
plot(pattern_random)
```

This could be achieved by turning the plot_* functions into plot.rd_pat and plot.rd_ras, check http://adv-r.had.co.nz/S3.html for instance.
