### 2021-10-20 r-spatialecology/shar#84 @lionel68
Axis in pair correlation plots
This is related to openjournals/joss-reviews#3811, the pair correlation plots from `plot_reconstructed_pattern` sometime show axis that do not span the whole range of the data:

```
pp <- spatstat.core::rLGCP(win=spatstat.geom::owin(c(0, 10), c(0, 10)))
ff <- fit_point_process(pp)
plot_reconstructed_pattern(ff)
```
![pcof](https://user-images.githubusercontent.com/7189850/138111789-b0231fae-e058-4c1b-9340-d948a8c87d2e.png)
