+++ R/reconstruct_pattern_marks.R
@@ -34,9 +34,7 @@
 #'
 #' @seealso
 #' \code{\link{fit_point_process}} \cr
-#' \code{\link{reconstruct_pattern_homo}} \cr
-#' \code{\link{reconstruct_pattern_hetero}} \cr
-#' \code{\link{reconstruct_pattern_cluster}}
+#' \code{\link{reconstruct_pattern}}
 #'
 #' @return rd_mar
 #'
