+++ R/reconstruct_pattern_homo.R
@@ -25,38 +25,6 @@
 #' @param verbose Logical if progress report is printed.
 #' @param plot Logical if pcf(r) function is plotted and updated during optimization.
 #'
-#' @details
-#' The functions randomizes the observed pattern by using pattern reconstruction
-#' as described in Tscheschel & Stoyan (2006) and Wiegand & Moloney (2014). The
-#' algorithm starts with a random pattern, shifts a point to a new location and
-#' keeps the change only, if the deviation between the observed and the reconstructed
-#' pattern decreases. The pair correlation function and the nearest neighbour
-#' distance function are used to describe the patterns.
-#'
-#' For large patterns (\code{n > comp_fast}) the pair correlation function can be estimated
-#' from Ripley's K-function without edge correction. This decreases the computational
-#' time. For more information see \code{\link{estimate_pcf_fast}}.
-#'
-#' The reconstruction can be stopped automatically if for n steps the energy does not
-#' decrease. The number of steps can be controlled by \code{no_change} and is set to
-#' \code{no_change = Inf} as default to never stop automatically.
-#'
-#' If \code{n_points} and \code{window} are not specified (default), the simulated pattern
-#' has the same number of points and window as the input pattern.
-#'
-#' The weights must be 0 < sum(weights) <= 1. To weight both summary functions identical,
-#' use \code{weights = c(0.5, 0.5)}.
-#'
-#' \code{spatstat} sets \code{r_length} to 513 by default. However, a lower value decreases
-#' the computational time while increasing the "bumpiness" of the summary function.
-#'
-#' @seealso
-#' \code{\link{calculate_energy}} \cr
-#' \code{\link{plot_randomized_pattern}} \cr
-#' \code{\link{reconstruct_pattern_hetero}} \cr
-#' \code{\link{reconstruct_pattern_cluster}} \cr
-#' \code{\link{reconstruct_pattern_marks}}
-#'
 #' @return rd_pat
 #'
 #' @examples
@@ -78,7 +46,7 @@
 #' Wiegand, T., & Moloney, K. A. (2014). Handbook of spatial point-pattern analysis
 #' in ecology. Boca Raton: Chapman and Hall/CRC Press.
 #'
-#' @export
+#' @keywords internal
 reconstruct_pattern_homo <- function(pattern,
                                      n_random = 1,
                                      e_threshold = 0.01,
