+++ tests/testthat/test-reconstruct_pattern_marks.R
@@ -1,57 +1,43 @@
 testthat::context("test-reconstruct_pattern_marks")
 
-pattern_recon <- reconstruct_pattern_homo(shar::species_a,
-                                          n_random = 1, return_input = FALSE,
+pattern_recon <- reconstruct_pattern_homo(species_a, n_random = 1, return_input = FALSE,
                                           simplify = TRUE, max_runs = 1,
                                           verbose = FALSE)
 
-marks_sub <- spatstat.geom::subset.ppp(shar::species_a, select = dbh)
+marks_sub <- spatstat.geom::subset.ppp(species_a, select = dbh)
 
 # normal reconstruction
-marks_recon <- shar::reconstruct_pattern_marks(pattern = pattern_recon,
-                                               marked_pattern = marks_sub,
-                                               n_random = 3,
-                                               max_runs = 1,
-                                               verbose = FALSE)
+marks_recon <- reconstruct_pattern_marks(pattern = pattern_recon, marked_pattern = marks_sub,
+                                         n_random = 3, max_runs = 1, verbose = FALSE)
 
 # no input
-marks_recon_ni <- shar::reconstruct_pattern_marks(pattern = pattern_recon,
-                                                  marked_pattern = marks_sub,
-                                                  n_random = 3,
-                                                  max_runs = 1,
-                                                  return_input = FALSE,
-                                                  verbose = FALSE)
+marks_recon_ni <- reconstruct_pattern_marks(pattern = pattern_recon, marked_pattern = marks_sub,
+                                            n_random = 3, max_runs = 1,
+                                            return_input = FALSE, verbose = FALSE)
 
 # simplify = TRUE
-marks_recon_simple <- shar::reconstruct_pattern_marks(pattern = pattern_recon,
-                                                      marked_pattern = marks_sub,
-                                                      n_random = 1,
-                                                      max_runs = 1,
-                                                      return_input = FALSE,
-                                                      simplify = TRUE,
-                                                      verbose = FALSE)
+marks_recon_simple <- reconstruct_pattern_marks(pattern = pattern_recon, marked_pattern = marks_sub,
+                                                n_random = 1, max_runs = 1,
+                                                return_input = FALSE, simplify = TRUE,
+                                                verbose = FALSE)
 
-marks_recon_energy <- shar::reconstruct_pattern_marks(pattern = pattern_recon,
-                                                      marked_pattern = marks_sub,
-                                                      n_random = 3,
-                                                      e_threshold = 0.1,
-                                                      verbose = FALSE)
+marks_recon_energy <- reconstruct_pattern_marks(pattern = pattern_recon, marked_pattern = marks_sub,
+                                                n_random = 3, e_threshold = 0.1,
+                                                verbose = FALSE)
 
 
 n_points_diff <- 100
 window_diff <- spatstat.geom::owin(xrange = c(0, 900), yrange = c(0, 1250))
 
-pattern_recon_diff <- reconstruct_pattern_homo(shar::species_a,
-                                               n_points = n_points_diff, window = window_diff,
-                                               n_random = 1, return_input = FALSE,
-                                               simplify = TRUE, max_runs = 1,
-                                               verbose = FALSE)
+pattern_recon_diff <- reconstruct_pattern_homo(species_a, n_points = n_points_diff,
+                                               window = window_diff, n_random = 1,
+                                               return_input = FALSE, simplify = TRUE,
+                                               max_runs = 1, verbose = FALSE)
 
-marks_recon_diff <- shar::reconstruct_pattern_marks(pattern = pattern_recon_diff,
-                                                    marked_pattern = marks_sub,
-                                                    n_random = 3,
-                                                    max_runs = 1,
[…]
