### 2021-10-27 r-spatialecology/shar#86 @Bisaloo
Use inherits() to test class inheritance
e.g.,

https://github.com/r-spatialecology/shar/blob/f4788a94fdd451d7d5aa86de2ced735f517b7f67/R/calculate_energy.R#L61

should be instead

```r
if (inherits(pattern, c("rd_pat", "rd_mar"))) {
```

Reference: https://developer.r-project.org/Blog/public/2019/11/09/when-you-think-class.-think-again/

This is part of https://github.com/openjournals/joss-reviews/issues/3811
