### 2021-11-01 d6516f3
#86 Use inherits()
