+++ R/plot_energy.R
@@ -12,11 +12,8 @@
 #' \code{reconstruct_pattern_*} .
 #'
 #' @seealso
-#' \code{\link{calculate_energy}} \cr
-#' \code{\link{reconstruct_pattern_homo}} \cr
-#' \code{\link{reconstruct_pattern_hetero}} \cr
-#' \code{\link{reconstruct_pattern_cluster}} \cr
-#' \code{\link{plot_randomized_pattern}}
+#' \code{\link{reconstruct_pattern}} \cr
+#' \code{\link{fit_point_process}}
 #'
 #' @return void
 #'
