+++ R/reconstruct_pattern.R
@@ -65,7 +65,6 @@
 #'
 #' @seealso
 #' \code{\link{calculate_energy}} \cr
-#' \code{\link{plot_randomized_pattern}} \cr
 #' \code{\link{reconstruct_pattern_marks}}
 #'
 #' @return rd_pat
