### 2021-11-01 27e4b4b
Some documentation improvements
