+++ R/results_habitat_association.R
@@ -17,11 +17,8 @@
 #' randomized count, respectively.
 #'
 #' @seealso
-#' \code{\link{randomize_raster}} \cr
-#' \code{\link{translate_raster}} \cr
-#' \code{\link{reconstruct_pattern_homo}} \cr
-#' \code{\link{reconstruct_pattern_hetero}} \cr
-#' \code{\link{reconstruct_pattern_cluster}}
+#' \code{\link{reconstruct_pattern}} \cr
+#' \code{\link{fit_point_process}}
 #'
 #' @return data.frame
 #'
