+++ R/calculate_energy.R
@@ -21,10 +21,8 @@
 #'
 #' @seealso
 #' \code{\link{plot_energy}} \cr
-#' \code{\link{reconstruct_pattern_homo}} \cr
-#' \code{\link{reconstruct_pattern_hetero}} \cr
-#' \code{\link{reconstruct_pattern_cluster}} \cr
-#' \code{\link{plot_randomized_pattern}}
+#' \code{\link{reconstruct_pattern}} \cr
+#' \code{\link{fit_point_process}}
 #'
 #' @return vector
 #'
