+++ R/plot.rd_mar.R
@@ -1,0 +1,234 @@
+#' plot.rd_mar
+#'
+#' @description Plot method for rd_pat object
+#'
+#' @param x rd_mar object with randomized patterns.
+#' @param what Character specifying to plot summary functions of point patterns
+#' (\code{what = "sf"}) or actual patterns (\code{what = "pp"}).
+#' @param probs Vector with quantiles of randomized data used for envelope construction.
+#' @param comp_fast Integer with threshold at which summary functions are estimated
+#' in a computational fast way.
+#' @param ask Logical if the user is asked to press <RETURN> before second summary function
+#' is plotted (only used if \code{what = "sf"}).
+#' @param verbose Logical if progress report is printed.
+#' @param ... Not used.
+#'
+#' @details
+#' The function plots the pair correlation function and the nearest neighbour function of
+#' the observed pattern and the reconstructed patterns (as "simulation envelopes").
+#' For large patterns \code{comp_fast = TRUE} decreases the computational demand because no edge
+#' correction is used and the pair correlation function is estimated based on Ripley's
+#' K-function. For more information see \code{\link{estimate_pcf_fast}}. It is also
+#' possible to plot 3 randomized patterns and the observed pattern using \code{what = "pp"}.
+#'
+#' @seealso
+#' \code{\link{reconstruct_pattern}} \cr
+#' \code{\link{fit_point_process}}
+#'
+#' @return void
+#'
+#' @examples
+#' \dontrun{
+#' pattern_recon <- reconstruct_pattern_homo(species_a, n_random = 1, max_runs = 1000,
+#' simplify = TRUE, return_input = FALSE)
+#' marks_sub <- spatstat.geom::subset.ppp(species_a, select = dbh)
+#' marks_recon <- reconstruct_pattern_marks(pattern_recon, marks_sub, n_random = 19, max_runs = 1000)
+#' plot(marks_recon)
+#' }
+#'
+#'
+#' @aliases plot.rd_mar
+#' @rdname plot.rd_mar
+
+#' @export
+plot.rd_mar <- function(x, what = "sf", probs = c(0.025, 0.975), comp_fast = 1000,
+                        ask = TRUE, verbose = TRUE, ...) {
+
+  # check if class is correct
+  if (!inherits(x = x, what = c("rd_pat", "rd_mar"))) {
+
+    stop("Class of 'pattern' must be 'rd_pat' or 'rd_mar'.", call. = FALSE)
+
+  }
+
+  # check if observed pattern is present
+  if (!spatstat.geom::is.ppp(x$observed)) {
+
+    stop("Input must include 'observed' pattern.", call. = FALSE)
+
+  }
+
+  pattern_names <- c(paste0("randomized_", seq(from = 1, to = length(x$randomized),
+                                               by = 1)), "observed")
+
+  if (what == "sf") {
+
+    # check if number of points exceed comp_fast limit
+    if (x$observed$n > comp_fast) {
+
+      comp_fast <- TRUE
+
+    } else {
+
+      comp_fast <- FALSE
+
+    }
+
+    name_unit <- spatstat.geom::unitname(x$observed)[[1]] # unit name for labels
+
+    # calculate r
+    r <- seq(from = 0, to = spatstat.core::rmax.rule(W = x$observed$window,
+                                                     lambda = spatstat.geom::intensity.ppp(x$observed)),
+             length.out = 250)
+
+    # combine observed and randomized to one list again
+    pattern <- c(x$randomized, list(x$observed))
+
+    names(pattern) <- pattern_names
+
+    result <- lapply(seq_along(pattern), function(x) {
+
+      mark_corr <- as.data.frame(spatstat.core::markcorr(pattern[[x]],
+                                                         correction = "Ripley",
+                                                         r = r))
+
+      # print progress
+      if (verbose) {
+
[…]
