+++ man/results_habitat_association.Rd
@@ -50,9 +50,6 @@ Species-area curves, spatial aggregation, and habitat specialization in tropical
 Journal of Theoretical Biology, 207(1), 81-99.
 }
 \seealso{
-\code{\link{randomize_raster}} \cr
-\code{\link{translate_raster}} \cr
-\code{\link{reconstruct_pattern_homo}} \cr
-\code{\link{reconstruct_pattern_hetero}} \cr
-\code{\link{reconstruct_pattern_cluster}}
+\code{\link{reconstruct_pattern}} \cr
+\code{\link{fit_point_process}}
 }
