+++ man/print.rd_pat.Rd
@@ -33,8 +33,6 @@ print(pattern_recon)
 
 }
 \seealso{
-\code{\link{reconstruct_pattern_homo}} \cr
-\code{\link{reconstruct_pattern_hetero}} \cr
-\code{\link{reconstruct_pattern_cluster}} \cr
+\code{\link{reconstruct_pattern}} \cr
 \code{\link{fit_point_process}}
 }

[…]
