### 2021-11-01 r-spatialecology/shar#81 @mhesselbarth
Fixed in d2fb2ae
