### 2021-11-01 68a5820
#84 Fix yrange in plotting
