### 2021-11-01 r-spatialecology/shar#84 @mhesselbarth
68a58204570ba046330760374ee9f2a508b696d4
Fixed. The y range is based on the min/max values now

``` r
library(shar)
library(spatstat)
pp <- spatstat.core::rLGCP(win=spatstat.geom::owin(c(0, 10), c(0, 10)))
ff <- fit_point_process(pp)
plot(ff, ask = FALSE)
```

![](https://i.imgur.com/VOqmJpZ.png)![](https://i.imgur.com/OsZwvN2.png)
