### 2021-11-01 45c3445
Update tests for pull request #85
