### 2021-11-03 6ee9c94
openjournals/joss-reviews#3811 Mention other point process and inlabru methods
