+++ paper.bib
@@ -1,3 +1,18 @@
+@article{Bachl2019,
+  title = {inlabru: an R package for Bayesian spatial modelling from ecological survey data},
+  shorttitle = {inlabru},
+  author = {Bachl, F.E. and Lindgren, F. and Borchers, D.L. and Illian, J.B.},
+  editor = {Freckleton, R.},
+  year = {2019},
+  month = jun,
+  journal = {Methods in Ecology and Evolution},
+  volume = {10},
+  number = {6},
+  pages = {760--766},
+  issn = {2041-210X, 2041-210X},
+  doi = {10.1111/2041-210X.13168},
+  langid = {english}
+}
 
 @book{Baddeley2015,
   title = {Spatial point patterns: Methodology and applications with R},
@@ -96,6 +111,17 @@
   issn = {09067590},
   doi = {10.1111/ecog.00911},
   langid = {english}
+}
+
+@article{Getzin2008,
+  title = {Heterogeneity influences spatial patterns and demographics in forest stands},
+  author = {Getzin, S. and Wiegand, T. and Wiegand, K. and He, F.},
+  year = {2008},
+  journal = {Journal of Ecology},
+  volume = {96},
+  number = {4},
+  pages = {807--820},
+  doi = {10.1111/j.1365-2745.2008.01377.x}
 }
 
 @article{Guo2016,
