### 2021-11-07 r-spatialecology/shar#87 @tretherington
Clarify expectations around NA data in the raster landscape
Regarding [https://github.com/openjournals/joss-reviews/issues/3811](https://github.com/openjournals/joss-reviews/issues/3811) I think there should be a note somewhere, perhaps in the README or the paper, maybe both around the expectation of how NA data in the raster landscape would be handled.

All the examples and supporting papers cited seem to work with landscapes that are "complete" in that points can occur anywhere within the unit-square for the point-process models and there is habitat everywhere within the raster landscape.  But I suspect that in reality it will be quite common for the study area to be a more complex shape than a rectangle, and therefore there will be NA cells within the raster landscape in which points cannot occur.  So I am left wondering if it is theoretically reasonable to use the methods in `shar` with that kind of data.  Essentially I am querying how NA data in a raster landscape is handled.

If `shar` can't handle NA data that is absolutely fine, but I do think it should be highlighted to potential users if that will preclude use of the package so that they can quickly eliminate `shar` as an option if it won't work for them.

Apologies if I have missed that information somewhere, but if it does exist, perhaps it could be made more prominent as I think that could have a massive bearing for potential users.
