### 2021-11-08 r-spatialecology/shar#87 @mhesselbarth
That is a very good point and I do agree this should be addressed. IMO, probably both in the readme as well as in the paper.

All methods with, exception of the torus translation, should be able "to handle" NAs in a way that they can be present under the prerequisite that a raster NA cell means that also no point could show up there (e..g think of a lake and tree locations). What is not allowed, that there are missing environmental data for cells of the landscape in which points can be present.
