### 2021-11-08 #3811 @mhesselbarth
Issues that still need to be addressed:

- [ ] r-spatialecology/shar#79
- [ ] r-spatialecology/shar#87
