### 2021-11-08 fc5c6cf
#87 Update NA value handling
