### 2021-11-08 1e84846
Update NA information in help files
