### 2021-11-08 r-spatialecology/shar#87 @mhesselbarth
fc5c6cf2d2590c5931d7289bfc7cc52db6c2650d , 1e8484691e6f81106927ec680bd8ba9a764c2647

Fixed. Please re-open if explanation is not sufficient enough.
