### 2021-11-08 r-spatialecology/shar#79 @mhesselbarth
> [...] So straight away I can see the counts are different, but these should be identical as these just reflect the counts of the species_a in the landscape both of which are fixed. [...]

Yes, the counts should be identical. I am not sure why they are not, I thought I fixed that. I will look into this, sorry.

> [...] The significance results are also confusing to me as while I do have the significant negative result for habitat 4 as the README describes, I also have significant positive relationships for habitat 2 and 3 [...]

This is partly due to the randomness of the methods, but not completely. It is also a "feature" of the methods. At one point, a negative association to one habitat has to lead to a positive association to another one (and vice versa). This is also  known and stated in the literature e.g. in Yamada et al. (2006): _"[...] If a species has positive association with one habitat, it necessarily has negative association with some others, and it is only a matter of the power of the test when it will become significant. [...]"_

I will add this explanation to the README and the help page of the data as well. Furthermore, I will look into the significance levels as suggested by you to ensure results are more "stable".

**References**
Yamada, T., Tomita, A., Itoh, A., Yamakura, T., Ohkubo, T., Kanzaki, M., Tan, S., Ashton, P.S., 2006. Habitat associations of Sterculiaceae trees in a Bornean rain forest plot. Journal of Vegetation Science 17, 559–566. [https://doi.org/10.1111/j.1654-1103.2006.tb02479.x](https://doi.org/10.1111/j.1654-1103.2006.tb02479.x)
