+++ paper.md
@@ -100,13 +100,13 @@ significance_level <- 0.01
 results_habitat_association(pattern = species_a, raster = random_walk, significance_level = significance_level)
 
 > Input: randomized raster
-> Quantile thresholds: negative < 0.025 || positive > 0.975
-habitat   count   lo    hi    significance
-1         9       2.95  13.15 n.s.
-2         25      8.95  25.30 n.s.
-3         27      11.00 23.20 positive
-4         0       13.95 27.30 negative
-5         12      6.00  17.05 n.s.
+> Quantile thresholds: negative < 0.005 || positive > 0.995
+  habitat count lo hi significance
+1       1    35 10 35         n.s.
+2       2    44 19 53         n.s.
+3       3    36 15 49         n.s.
+4       4     4 15 58     negative
+5       5    73 48 90         n.s.
 ```
 
 The `shar` packages also provides several utility and plotting functions such as `plot_randomized_raster()` and `plot_randomized_pattern()` to plot the null model data, `calculate_energy()` to calculate the difference between the input object locations and the randomized null model data object locations, or `classify_habitats()` to classify continuous environmental data into discrete habitats.  
