### 2021-11-08 r-spatialecology/shar#79 @mhesselbarth
README and paper should be updated with the correct example data now!
