### 2021-11-08 r-spatialecology/shar#79 @tretherington
This all looks great, many thanks. 😄

However, it looks like you've renamed or removed the `reconstruct_pattern_cluster` function that is still used in the README examples.  I used the `reconstruct_pattern` function instead which works, but you should update the examples to refer to the new function name.
