### 2021-11-08 r-spatialecology/shar#80 @tretherington
Thanks for expanding on these points, and having now skimmed through the cited papers I think you have emphasised the importance of the spatial patterning within the points, and why using a purely random approach is not entirely useful.  I have not considered this before, so this has been a useful concept for me to learn, so I appreciate the effort you have gone to here.
