### 2021-11-09 155c008
Fix README reconstruction function
