+++ README.Rmd
@@ -79,7 +79,7 @@ To randomize the point pattern, either use the Gamma test described by Plotkin e
 gamma_test <- fit_point_process(pattern = species_b, process = "cluster", n_random = 99)
 
 # (this can takes some time)
-reconstruction <- reconstruct_pattern_cluster(pattern = species_b, n_random = 99, e_threshold = 0.05)
+reconstruction <- reconstruct_pattern(pattern = species_b, n_random = 99, e_threshold = 0.05)
 ```
 
 Of course, there are several utility functions. For example, you can plot a randomized pattern or calculate the differences between the observed pattern and the randomized patterns (using summary functions). 
