### 2021-11-09 r-spatialecology/shar#79 @mhesselbarth
Glad to hear things look better now.

Yes, `reconstruct_pattern` is a new wrapper I introduced. Should be fixed in 155c008.
