### 2021-11-09 #3811 @mhesselbarth
@Bisaloo
I am planning to use v1.3 for the next release with all the helpful comments and suggestions by the reviewers. The first @whedon comment in this thread says v1.2.2. Not sure if thats important.
