### 2021-11-09 fab015d
:pencil: Remove one row from badges table
