### 2021-11-09 #3811 @tretherington
I have just tried to run over all the points again, and for me I think the only outstanding one might be: "State of the field: Do the authors describe how this software compares to other commonly-used packages?"

At the moment the paper doesn't refer to any other packages that might be commonly used to do species-habitat analyses.  I acknowledge that the package is quite possibly unique in that it does habitat analyses that account for the spatial patterns in the occurrence and habitat data, as opposed to more classical methods that just bin the data and assume randomness.  But I do wonder if it would be helpful to make reference to some commonly used R packages that apply these more simplistic (perhaps incorrect!?) techniques as I think it could help a potential user to create a cognitive connection between how they currently do things, and how they could do things better by using `shar`.

For example, if I were to read that a function in a package I commonly use might be inappropriate then I think that is more likely to grab my attention as that provides a massive indicator of relevance.  So perhaps at the end of the first paragraph of the statement of need section in the paper where you describe the sub-optimal approach you could give some examples of packages that include that sub-optimal functionality?

The problem I have here is that the species distribution/ecological niche modelling I tend to do is not really comparable to this approach - as I use continuous variables to produce a continuous model rather than a table of results from categorical data.  So I'm unable to readily identify R packages that do similar habitat analyses.  From vague memory I think the `adehabitatHS` package might be a good example as it includes a `compana` function for compositional analysis of habitat use that I think would qualify as an example of an approach that assumes statistical independence of the observations.  But after that I am stuck, but perhaps you will have a better idea given you have more experience in this area.

Apologies that this isn't a very specific comment 😬 but hopefully that makes some sense?
