### 2021-11-10 #3811 @Bisaloo
> I am planning to use v1.3 for the next release with all the helpful comments and suggestions by the reviewers. The first @whedon comment in this thread says v1.2.2. Not sure if thats important.

That's not a problem. Feel free to increment the version number. We ask you to do it right before acceptance anyways. I'll tell whedon what version should be marked as accepted later.
