+++ paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 Studying species-habitat associations is one approach to reveal the importance of abiotic processes in shaping the spatial distribution of species.
 Even though the `R` programming language offers many packages for spatial point pattern analysis, currently there is no comprehensive package specifically designed to analyze species-habitat associations.
-The `shar` package builds on widely used `R` packages for spatial analyses and provides an easy and straightforward way to uncover species-habitat associations.
+The `shar` package builds on widely used `R` packages for spatial analyses and provides an easy and straightforward way to uncover species-habitat associations for discrete environmental data.
 
 # Statement of need
 
@@ -37,13 +37,16 @@ Previous research has shown that violating the independence assumptions resulted
 
 Ecologists use the spatial distribution of ecological objects to infer the underlying processes that shaped their distribution because spatial patterns can act as a "memory" of the processes that shaped them [@Velazquez2016].
 A spatial point pattern contains the fully mapped locations (in terms of *x*,*y* coordinates) of all individual objects in a normally two-dimensional study area and assumes that the object locations can be described by discrete points [@Wiegand2014; @Velazquez2016].
-For example, many studies use individual tree locations to infer the processes that determined their distribution [@Velazquez2016], and further examples include gopher mounds [@Klaas2000], gorilla nests [@Funwi-Gabga2012], "fairy circles" [@Getzin2015], or bacteria on leaves [@Esser2015].
-The `spatstat` package [@Baddeley2015] allows ecologists to access many methods of spatial point pattern analysis, such as summary functions and null model simulations.
-However, even though many ecological studies on species-habitat associations can be found in the literature [@John2007; @Garzon-Lopez2014; @Guo2016; @Yang2016; @Du2017; @Furniss2017], `spatstat` cannot be used to reveal species-habitat associations without larger programming efforts by the users.
+For example, many studies use individual tree locations to infer the processes that determined their distribution [@Velazquez2016], and further examples include gopher mounds [@Klaas2000], gorilla nests [@Funwi-Gabga2012], "fairy circles" [@Getzin2015], or bacteria on leaves [@Esser2015a].
 
-Thus, the `shar` package was developed to provide a tool set to analyze species-habitat associations of spatial point patterns.
+The `spatstat` package [@Baddeley2015] allows ecologists to access many methods of spatial point pattern analysis, such as summary functions and null model simulations, and to simulate heterogeneous point process models that can be used to show the importance of abiotic processes for continuous environmental data [@Getzin2008].
+However, even though many ecological studies on species-habitat associations using discrete environmental data can be found in the literature [@John2007; @Garzon-Lopez2014; @Guo2016; @Yang2016; @Du2017; @Furniss2017], `spatstat` cannot be used to reveal such associations without larger programming efforts by the users.
+The `inlabru` package provides an approach to  analyze the importance of abiotic processes, mostly for continuous environmental data, using Bayesian spatial modelling (i.e., `inlabru`) [@Bachl2019].
+The `fgeo` package [@Lepore2019] allows to visualize and analyze forest diversity, including species-habitat associations.
+But, the `fgeo` was designed to specifally handle ForestGEO data (<https://forestgeo.si.edu>) and furthermore includes only a small subset of methods available to analyze species-habitat associations.
+
+Thus, the `shar` package was developed to provide a comprehensive tool set to analyze species-habitat associations of spatial point patterns.
 All methods in the `shar` package are designed for discrete environmental data and have the advantage of very few assumptions about the data characteristics.
-Nevertheless, also other methods exist to analyze the importance of abiotic processes, especially for continuous environmental data, such as heterogeneous point process models [@Getzin2008] or Bayesian spatial modelling [@Bachl2019].
[…]
