+++ paper.bib
@@ -1,25 +1,11 @@
-@article{Bachl2019,
-  title = {inlabru: an R package for Bayesian spatial modelling from ecological survey data},
-  shorttitle = {inlabru},
-  author = {Bachl, F.E. and Lindgren, F. and Borchers, D.L. and Illian, J.B.},
-  editor = {Freckleton, R.},
-  year = {2019},
-  month = jun,
-  journal = {Methods in Ecology and Evolution},
-  volume = {10},
-  number = {6},
-  pages = {760--766},
-  issn = {2041-210X, 2041-210X},
-  doi = {10.1111/2041-210X.13168},
-  langid = {english}
-}
 
 @book{Baddeley2015,
   title = {Spatial point patterns: Methodology and applications with R},
   author = {Baddeley, A. and Rubak, E. and Turner, R.},
   year = {2015},
   publisher = {Chapman and Hall/CRC Press},
-  address = {London}
+  address = {London, UK},
+  isbn = {978-1-4822-1020-0}
 }
 
 @article{Bengtsson2021,
@@ -52,7 +38,7 @@
   doi = {10.1038/s41598-017-15789-5}
 }
 
-@article{Esser2015,
+@article{Esser2015a,
   title = {Spatial scales of interactions among bacteria and between bacteria and the leaf surface},
   author = {Esser, D.S. and Leveau, J.H.J. and Meyer, K.M. and Wiegand, K.},
   year = {2015},
@@ -103,25 +89,11 @@
   title = {Adopting a spatially explicit perspective to study the mysterious fairy circles of Namibia},
   author = {Getzin, S. and Wiegand, K. and Wiegand, T. and Yizhaq, H. and {von Hardenberg}, J. and Meron, E.},
   year = {2015},
-  month = jan,
   journal = {Ecography},
   volume = {38},
   number = {1},
   pages = {1--11},
-  issn = {09067590},
-  doi = {10.1111/ecog.00911},
-  langid = {english}
-}
-
-@article{Getzin2008,
-  title = {Heterogeneity influences spatial patterns and demographics in forest stands},
-  author = {Getzin, S. and Wiegand, T. and Wiegand, K. and He, F.},
-  year = {2008},
-  journal = {Journal of Ecology},
-  volume = {96},
-  number = {4},
-  pages = {807--820},
-  doi = {10.1111/j.1365-2745.2008.01377.x}
+  doi = {10.1111/ecog.00911}
 }
 
 @article{Guo2016,
@@ -163,16 +135,6 @@
   doi = {10.1073/pnas.0604666104}
 }
 
-@article{Klaas2000,
-  title = {The tempo and mode of gopher mound production in a tallgrass prairie remnant},
-  author = {Klaas, B.A. and Moloney, K.A. and Danielson, B.J.},
-  year = {2000},
-  journal = {Ecography},
-  volume = {23},
-  pages = {246--256},
-  doi = {10.1111/j.1600-0587.2000.tb00280.x}
-}
-
 @article{Kirkpatrick1983,
   title = {Optimization by simulated annealing},
   author = {Kirkpatrick, S. and Gelatt, C.D.Jr. and Vecchi, M.P.},
@@ -184,6 +146,22 @@
   doi = {10.1126/science.220.4598.671}
 }
 
+@article{Klaas2000,
+  title = {The tempo and mode of gopher mound production in a tallgrass prairie remnant},
+  author = {Klaas, B.A. and Moloney, K.A. and Danielson, B.J.},
+  year = {2000},
+  journal = {Ecography},
+  volume = {23},
+  pages = {246--256},
+  doi = {10.1111/j.1600-0587.2000.tb00280.x}
+}
+
[…]
