+++ paper.bib
@@ -1,4 +1,20 @@
 
+@article{Bachl2019,
+  title = {inlabru: an R package for Bayesian spatial modelling from ecological survey data},
+  shorttitle = {inlabru},
+  author = {Bachl, F.E. and Lindgren, F. and Borchers, D.L. and Illian, J.B.},
+  editor = {Freckleton, R.},
+  year = {2019},
+  month = jun,
+  journal = {Methods in Ecology and Evolution},
+  volume = {10},
+  number = {6},
+  pages = {760--766},
+  issn = {2041-210X, 2041-210X},
+  doi = {10.1111/2041-210X.13168},
+  langid = {english}
+}
+
 @book{Baddeley2015,
   title = {Spatial point patterns: Methodology and applications with R},
   author = {Baddeley, A. and Rubak, E. and Turner, R.},
@@ -83,6 +99,17 @@
   number = {2},
   pages = {349--362},
   doi = {10.1111/jvs.12090}
+}
+
+@article{Getzin2008,
+  title = {Heterogeneity influences spatial patterns and demographics in forest stands},
+  author = {Getzin, S. and Wiegand, T. and Wiegand, K. and He, F.},
+  year = {2008},
+  journal = {Journal of Ecology},
+  volume = {96},
+  number = {4},
+  pages = {807--820},
+  doi = {10.1111/j.1365-2745.2008.01377.x}
 }
 
 @article{Getzin2015,
@@ -171,6 +198,14 @@
   number = {1},
   pages = {81--99},
   doi = {10.1006/jtbi.2000.2158}
+}
+
+@misc{RCoreTeam2021,
+  title = {R: A language and environment for statistical computing.},
+  author = {{R Core Team}},
+  year = {2021},
+  address = {Vienna (Austria)},
+  howpublished = {R Foundation for Statistical Computing}
 }
 
 @incollection{Tilman1993,
