### 2021-11-10 #3811 @mhesselbarth
@tretherington
https://github.com/r-spatialecology/shar/commit/3850fb113935544b61fa033e637cba45e0ebb365

I changed the last 1,2 paragraphs of the statement of need section quite a bit trying to compare `shar` more to other existing packages and methods.

However, I don't think that including `adehabitatHS` makes too much sense, since it is not really based on a spatially explicit point pattern approach, but rather on matrices of habitat use.
