+++ vignettes/get_started.Rmd
@@ -18,10 +18,7 @@ knitr::opts_chunk$set(
 ```
 
 ```{r load_libraries_hidden, eval = TRUE, echo = FALSE, message = FALSE, results = "hide"}
-library(dplyr)
 library(shar)
-library(spatstat.core)
-library(raster)
 ```
 
 ## Design
