+++ .github/workflows/Deploy-pkgdown.yaml
@@ -23,7 +23,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
-          extra-packages: pkgdown
+          extra-packages: pkgdown, future, future.apply
           needs: website
 
       - name: Deploy package
