+++ .github/workflows/R-CMD-check.yaml
@@ -13,7 +13,8 @@ on:
 name: R-CMD-check
 
 jobs:
-  R-CMD-check:
+  r-cmd-check:
+    if: "!contains(github.event.commits[0].message, '[skip-ci]')"
     runs-on: ${{ matrix.config.os }}
 
     name: ${{ matrix.config.os }} (${{ matrix.config.r }})
