### 2021-11-11 #3811 @tretherington
Thanks for this.  I understand that there may not be anything comparable in terms of methods to `shar`, but I guess I was thinking that if everyone is using a package that uses the equivalent suboptimal methods, then pointing out that limitation would be helpful.  But you are the subject expert here, so I'm happy to go with what you have done. 😄
