### 2021-11-11 #3811 @tretherington
@Bisaloo I've now ticked off everything in my reviewer checklist 🥳 but as a first time reviewer for JOSS if I've missed a step please let me know.
