### 2021-11-11 #3811 @Bisaloo
> @Bisaloo I've now ticked off everything in my reviewer checklist partying_face but as a first time reviewer for JOSS if I've missed a step please let me know.

@tretherington, you're all good. Thank you so much for your very helpful comments! :100:
