### 2021-11-11 #3811 @Bisaloo
@lionel68, could you please go over you checklist in [this comment](https://github.com/openjournals/joss-reviews/issues/3811#issue-1022291832)?
After reviewing each item, either check it or open an issue in https://github.com/r-spatialecology/shar/issues if you think it should be improved. Thanks :pray:
