+++ .github/workflows/Render-README.yaml
@@ -1,32 +1,37 @@
+# Workflow derived from https://github.com/r-lib/actions/tree/master/examples
+# Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    paths:
-      - README.Rmd
+    paths: ['README.Rmd']
 
 name: Render-README
 
 jobs:
-  render:
-    name: Render README
-    runs-on: macOS-latest
+  render-readme:
+    runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
-      - uses: actions/checkout@v2
-      - uses: r-lib/actions/setup-r@v1
+      - name: Checkout repo
+        uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+
       - uses: r-lib/actions/setup-pandoc@v1
-      - name: Install rmarkdown, remotes, and the local package
-        run: |
-          install.packages("remotes")
-          remotes::install_local(".")
-          remotes::install_cran("rmarkdown")
-        shell: Rscript {0}
-      - name: Render README
+
+      - uses: r-lib/actions/setup-r@v1
+
+      - uses: r-lib/actions/setup-r-dependencies@v1
+        with:
+          extra-packages: rmarkdown
+
+      - name: Render README files
         run: Rscript -e 'rmarkdown::render("README.Rmd")'
-      - name: Commit results
+
+      - name: Commit README
         run: |
           git config --local user.email "actions@github.com"
           git config --local user.name "GitHub Actions"
-          git add README.md man/figures/README-*
-          git commit -m 'Re-build README.Rmd' || echo "No changes to commit"
+          git add README.md
+          git commit -m "Re-build README.md" || echo "No changes to commit"
           git push origin || echo "No changes to commit"
