+++ .github/workflows/Test-coverage.yaml
@@ -2,9 +2,9 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, development]
+    branches: [CRAN, main, development]
   pull_request:
-    branches: [main, development]
+    branches: [CRAN, main]
 
 name: Test-coverage
 
