### 2021-11-19 bcb0beb
:white_check_mark: Try to install GDAL in GH action
