+++ DESCRIPTION
@@ -16,7 +16,7 @@ Description:
 License: GPL (>= 3)
 URL: https://r-spatialecology.github.io/shar/
 BugReports: https://github.com/r-spatialecology/shar/issues/
-Depends: R (>= 3.2.0)
+Depends: R (>= 3.1.0)
 Imports:
     classInt,
     graphics,
