+++ .github/workflows/R-CMD-check.yaml
@@ -39,6 +39,10 @@ jobs:
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
 
+      - name: Install GDAL
+        if: runner.os == 'macOS'
+        run: brew install gdal
+
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
           extra-packages: rcmdcheck
