### 2021-11-19 b4b011c
Try to fix RCMDCHECK
