+++ .github/workflows/R-CMD-check.yaml
@@ -39,9 +39,11 @@ jobs:
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
 
-      - name: Install GDAL
+      - name: Install spatial dependencie
         if: runner.os == 'macOS'
-        run: brew install gdal
+        run: |
+          brew install gdal
+          brew install proj
 
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
