### 2021-11-19 a0b2362
Install spatial deps before R
