### 2021-11-23 #3811 @lionel68
Thanks @whedon

Some further / last minor comments:

1. line 24: I would rephrase, "Spatial autocorrelation of the environmental conditions could also ..."
2. line 27: this sentence is a bit cryptic regarding the method developed in the package, maybe best to directly name it? Like "Previous research have shown that violation of independence assumptions led to higher number of false-positive habitat associations when using method XX compared to method YY"
3. line 104: didn't function `reconstruct_pattern` changed in previous commit [beec06e](https://github.com/r-spatialecology/shar/commit/beec06e422a22088450cceacfe4d62871e438d4a) ?

Otherwise looks good!
