### 2021-11-23 #3811 @mhesselbarth
- [ ]  line 24: I would rephrase, "Spatial autocorrelation of the environmental conditions could also ..."
- [ ]  line 27: this sentence is a bit cryptic regarding the method developed in the package, maybe best to directly name it? Like "Previous research have shown that violation of independence assumptions led to higher number of false-positive habitat associations when using method XX compared to method YY"
- [ ]  line 104: didn't function reconstruct_pattern changed in previous commit beec06e ?
