### 2021-11-23 75c571b
openjournals/joss-reviews#3811 Some more improvements of paper.md
