### 2021-11-23 #3811 @mhesselbarth
Fixed in r-spatialecology/shar@75c571b188d03ec82fcb48b6bc5f3ab23fe28063
