### 2021-11-23 34332d3
:lipstick: Update citation and some cosmetic changes
