+++ man/species_a.Rd
@@ -19,6 +19,5 @@ associations to other habitats (Yamada et al. 2006).
 Yamada, T., Tomita, A., Itoh, A., Yamakura, T., Ohkubo, T., Kanzaki, M., Tan, S.,
 Ashton, P.S., 2006. Habitat associations of Sterculiaceae trees in a Bornean rain
 forest plot. Journal of Vegetation Science 17, 559–566.
-https://doi.org/10.1111/j.1654-1103.2006.tb02479.x
 }
 \keyword{datasets}
