### 2021-11-23 #3811 @Bisaloo
@lionel68, you didn't check "State of the field: Do the authors describe how this software compares to other commonly-used packages?".

Is this on purpose or did you miss it?
