+++ DESCRIPTION
@@ -4,7 +4,7 @@ Title: Species-Habitat Associations
 Version: 1.2.2
 Authors@R: c(person("Maximillian H.K.", "Hesselbarth", email = "mhk.hesselbarth@gmail.com",
              role = c("aut", "cre"), comment = c(ORCID = "0000-0003-1125-9918")),
-             person("Marco", "Sciaini", email = "sciaini.marco@gmail.com",
+             person("Marco", "Sciaini", email = "marco.sciaini@posteo.net",
              role = "aut", comment = c(ORCID = '0000-0002-3042-5435')))
 Maintainer: Maximillian H.K. Hesselbarth <mhk.hesselbarth@gmail.com>
 Description:
