### 2021-12-01 1c41ea7
:lipstick: Update some examples related to reconstruct_pattern
