+++ R/translate_raster.R
@@ -28,7 +28,8 @@
 #' landscape_classified <- classify_habitats(landscape, classes = 5)
 #'
 #' landscape_random <- translate_raster(landscape_classified)
-#' landscape_random_sub <- translate_raster(landscape_classified, steps_x = 1:10, steps_y = 1:5)
+#' landscape_random_sub <- translate_raster(landscape_classified,
+#' steps_x = 1:10, steps_y = 1:5)
 #' }
 #'
 #' @aliases translate_raster
