+++ R/plot.rd_pat.R
@@ -32,7 +32,8 @@
 #' pattern_random <- fit_point_process(species_a, n_random = 39)
 #' plot(pattern_random)
 #'
-#' pattern_recon <- reconstruct_pattern_hetero(species_b, n_random = 19, max_runs = 1000)
+#' pattern_recon <- reconstruct_pattern(species_b, n_random = 19,
+#' max_runs = 1000, method = "hetero")
 #' plot(pattern_recon)
 #' }
 #'
