+++ R/plot.rd_mar.R
@@ -32,7 +32,8 @@
 #' pattern_recon <- reconstruct_pattern(species_a, n_random = 1, max_runs = 1000,
 #' simplify = TRUE, return_input = FALSE)
 #' marks_sub <- spatstat.geom::subset.ppp(species_a, select = dbh)
-#' marks_recon <- reconstruct_pattern_marks(pattern_recon, marks_sub, n_random = 19, max_runs = 1000)
+#' marks_recon <- reconstruct_pattern_marks(pattern_recon, marks_sub,
+#' n_random = 19, max_runs = 1000)
 #' plot(marks_recon)
 #' }
 #'
