+++ R/print.rd_pat.R
@@ -20,7 +20,8 @@
 #' print(pattern_random)
 #'
 #' \dontrun{
-#' pattern_recon <- reconstruct_pattern_hetero(species_b, n_random = 19, max_runs = 1000)
+#' pattern_recon <- reconstruct_pattern(species_b, n_random = 19, max_runs = 1000,
+#' method = "hetero")
 #' print(pattern_recon)
 #' }
 #'
