+++ man/list_to_randomized.Rd
@@ -36,7 +36,5 @@ list_to_randomized(list = fit_list, observed = species_a)
 \seealso{
 \code{\link{randomize_raster}} \cr
 \code{\link{translate_raster}} \cr
-\code{\link{reconstruct_pattern_homo}} \cr
-\code{\link{reconstruct_pattern_hetero}} \cr
-\code{\link{reconstruct_pattern_cluster}}
+\code{\link{reconstruct_pattern}}
 }
