+++ _pkgdown.yml
@@ -1,3 +1,5 @@
+url: https://r-spatialecology.github.io/shar
+
 title: shar
 
 template:
