### 2021-12-01 3cfda8c
[skip-ci] Adding ../man/figures/ to articles
