+++ _pkgdown.yml
@@ -1,10 +1,8 @@
-url: https://r-spatialecology.github.io/shar
-
-title: shar
+url: https://r-spatialecology.github.io/shar/
 
 template:
-  params:
-    bootswatch: cosmo
+  bootstrap: 5
+  bootswatch: minty
 
 reference:
   - title: Data sets
@@ -41,7 +39,6 @@ reference:
       - print.rd_ras
 
 navbar:
-  title: shar
   left:
     - text: "Functions"
       icon: fa-gears
@@ -49,21 +46,21 @@ navbar:
     - text: "Get Started"
       icon: fa fa-magic
       href: articles/get_started.html
+    - text: "General Background"
+      icon: fa-book
+      href: articles/background.html
     - text: "More"
       icon: fa-graduation-cap
       menu:
-       - text: "General Background"
-         icon: fa-book
-         href: articles/articles/background.html
-       - text: "Reconstruct multiple pattern"
-         icon: fa-book
-         href: articles/articles/reconstruct_multiple_patterns.html
-       - text: "Parallelization"
-         icon: fa-book
-         href: articles/articles/parallelization.html
-       - text: "Publication Record"
-         icon: fa-file-text-o
-         href: articles/articles/publication_record.html
+        - text: "Reconstruct multiple pattern"
+          icon: fa-book
+          href: articles/reconstruct_multiple_patterns.html
+        - text: "Parallelization"
+          icon: fa-book
+          href: articles/parallelization.html
+        - text: "Publication Record"
+          icon: fa-file-text-o
+          href: articles/publication_record.html
   right:
     - icon: fa-newspaper-o
       text: "News"
@@ -74,4 +71,3 @@ navbar:
     - icon: fa-user
       text: "Contact"
       hreft: https://mhesselbarth.rbind.io
-
