+++ README.Rmd
@@ -6,8 +6,8 @@ output: github_document
 
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
-  collapse = TRUE,
-  fig.path = "man/figures/README-"
+  collapse = TRUE, 
+  fig.path = "man/figures/"
 )
 ```
 
