+++ README.md
@@ -94,8 +94,6 @@ col = c("#440154FF", "#3B528BFF", "#21908CFF", "#5DC863FF", "#FDE725FF")
 plot(torus_trans, n = 3, col = col)
 ```
 
-<img src="man/figures/plot_habitat-random-1.png" style="display: block; margin: auto;" />
-
 To randomize the point pattern, either use the Gamma test described by
 Plotkin et al. (2000) or pattern reconstruction (Kirkpatrick et
 al. 1983; Tscheschel & Stoyan 2006).
@@ -115,50 +113,8 @@ observed pattern and the randomized patterns (using summary functions).
 plot(reconstruction, verbose = FALSE, ask = FALSE)
 ```
 
-<img src="man/figures/plot-random_pattern-1.png" style="display: block; margin: auto;" /><img src="man/figures/plot-random_pattern-2.png" style="display: block; margin: auto;" />
-
 ``` r
 calculate_energy(reconstruction, verbose = FALSE)
-##  randomized_1  randomized_2  randomized_3  randomized_4  randomized_5 
-##    0.04977170    0.04957230    0.04726960    0.04850374    0.04943275 
-##  randomized_6  randomized_7  randomized_8  randomized_9 randomized_10 
-##    0.04961647    0.04956624    0.04899253    0.04990957    0.04986367 
-## randomized_11 randomized_12 randomized_13 randomized_14 randomized_15 
-##    0.04950447    0.04874728    0.04929699    0.04946615    0.04920561 
-## randomized_16 randomized_17 randomized_18 randomized_19 randomized_20 
-##    0.04990415    0.04813098    0.04967062    0.04964872    0.04990442 
-## randomized_21 randomized_22 randomized_23 randomized_24 randomized_25 
-##    0.04962960    0.04897384    0.04933126    0.04793454    0.04684534 
-## randomized_26 randomized_27 randomized_28 randomized_29 randomized_30 
-##    0.04937752    0.04901977    0.04727432    0.04872280    0.04909555 
-## randomized_31 randomized_32 randomized_33 randomized_34 randomized_35 
-##    0.04943618    0.04911176    0.04961079    0.04722192    0.04903119 
-## randomized_36 randomized_37 randomized_38 randomized_39 randomized_40 
-##    0.04871886    0.04850467    0.04816963    0.04959528    0.04882913 
-## randomized_41 randomized_42 randomized_43 randomized_44 randomized_45 
-##    0.04992851    0.04971571    0.04823854    0.04926482    0.04919277 
-## randomized_46 randomized_47 randomized_48 randomized_49 randomized_50 
-##    0.04878647    0.04944433    0.04970445    0.04936347    0.04940492 
-## randomized_51 randomized_52 randomized_53 randomized_54 randomized_55 
-##    0.04928191    0.04990232    0.04821732    0.04914895    0.04945262 
-## randomized_56 randomized_57 randomized_58 randomized_59 randomized_60 
-##    0.04935535    0.04945049    0.04945263    0.04968753    0.04908163 
-## randomized_61 randomized_62 randomized_63 randomized_64 randomized_65 
-##    0.04894439    0.04965690    0.04674924    0.04953226    0.04954639 
-## randomized_66 randomized_67 randomized_68 randomized_69 randomized_70 
-##    0.04786011    0.04967592    0.04798818    0.04918426    0.04903230 
-## randomized_71 randomized_72 randomized_73 randomized_74 randomized_75 
-##    0.04844243    0.04979140    0.04890577    0.04944504    0.04892242 
-## randomized_76 randomized_77 randomized_78 randomized_79 randomized_80 
-##    0.04951001    0.04983087    0.04982151    0.04949764    0.04996596 
[…]
