+++ README.Rmd
@@ -67,10 +67,12 @@ torus_trans <- translate_raster(raster = landscape_classified)
 random_walk <- randomize_raster(raster = landscape_classified, n_random = 99)
 ```
 
-```{r plot_habitat-random, eval = FALSE}
-col = c("#440154FF", "#3B528BFF", "#21908CFF", "#5DC863FF", "#FDE725FF")
+To plot the randomized raster, you can use the plot function and specify the number of raster as as well as the color palette used for the discrete environmental data.
 
-plot(torus_trans, n = 3, col = col)
+```{r plot_habitat-random, eval = FALSE}
+col_palette <- c("#440154FF", "#3B528BFF", "#21908CFF", "#5DC863FF", "#FDE725FF")
+
+plot(torus_trans, n = 3, col = col_palette)
 ```
 
 To randomize the point pattern, either use the Gamma test described by Plotkin et al. (2000) or pattern reconstruction (Kirkpatrick et al. 1983; Tscheschel & Stoyan 2006). 
@@ -82,11 +84,13 @@ gamma_test <- fit_point_process(pattern = species_b, process = "cluster", n_rand
 reconstruction <- reconstruct_pattern(pattern = species_b, n_random = 99, e_threshold = 0.05)
 ```
 
-Of course, there are several utility functions. For example, you can plot a randomized pattern or calculate the differences between the observed pattern and the randomized patterns (using summary functions). 
+Of course, there are several utility functions. For example, you can plot the summary function of the observed pattern and the simulation envelopes of randomized patterns using the plot function. 
 
 ```{r plot-random_pattern, eval = FALSE}
-plot(reconstruction, verbose = FALSE, ask = FALSE)
+plot(reconstruction)
 ```
+
+Another utility functions allows to calculate the differences between the observed pattern and the randomized patterns (also called energy using summary functions). 
 
 ```{r calculate-energy, eval = FALSE}
 calculate_energy(reconstruction, verbose = FALSE)
@@ -131,4 +135,3 @@ Tscheschel, A., Stoyan, D., 2006. Statistical reconstruction of random point pat
 Wiegand, T., Moloney, K.A., 2014. Handbook of spatial point-pattern analysis in ecology. Chapman and Hall/CRC Press, Boca Raton. <isbn:9781420082548>
 
 Yamada, T., Tomita, A., Itoh, A., Yamakura, T., Ohkubo, T., Kanzaki, M., Tan, S., Ashton, P.S., 2006. Habitat associations of Sterculiaceae trees in a Bornean rain forest plot. Journal of Vegetation Science 17, 559–566. <https://doi.org/10.1111/j.1654-1103.2006.tb02479.x>
-
