### 2021-12-01 e8baac7
[skip-ci] Remove logo from README for now...
