+++ README.Rmd
@@ -20,7 +20,7 @@ knitr::opts_chunk$set(
 
 <!-- badges: end -->
 
-<img src="man/figures/logo.png" align="right" width="150" />
+<!-- <img src="man/figures/logo.png" align="right" width="150" /> -->
 
 # shar
 
