+++ README.md
@@ -10,7 +10,7 @@
 
 <!-- badges: end -->
 
-<img src="man/figures/logo.png" align="right" width="150" />
+<!-- <img src="man/figures/logo.png" align="right" width="150" /> -->
 
 # shar
 
