### 2021-12-01 528237b
[skip-ci] Remove logo completely
