+++ README.Rmd
@@ -20,8 +20,6 @@ knitr::opts_chunk$set(
 
 <!-- badges: end -->
 
-<!-- <img src="man/figures/logo.png" align="right" width="150" /> -->
-
 # shar
 
 **S**pecies-**H**abitat **A**ssociations in **R** is a `R` package to analyze species-habitat associations. Therefore, information about the location of the species is needed (as a point pattern) and about the environmental conditions (as a raster map). In order to analyse the data for significant habitat associations either the location data or the environmental data is randomized n-times. Then, counts within the habitats are compared between the randomized data and the observed data. Positive or negative associations are present if the observed counts is higher or lower than the randomized counts (using quantile thresholds). Methods are mainly described in Plotkin et al. (2000), Harms et al. (2001) and Wiegand & Moloney (2014). **shar** is mainly based on the [`spatstat`](http://spatstat.org) (Baddeley et al. 2015) and [`raster`](https://rspatial.org/raster/) (Hijmans 2017) package.
