+++ pkgdown.yml
@@ -1,0 +1,14 @@
+pandoc: 2.7.3
+pkgdown: 2.0.0
+pkgdown_sha: ~
+articles:
+  background: background.html
+  parallelization: parallelization.html
+  publication_record: publication_record.html
+  reconstruct_multiple_patterns: reconstruct_multiple_patterns.html
+  get_started: get_started.html
+last_built: 2021-12-01T20:50Z
+urls:
+  reference: https://r-spatialecology.github.io/shar/reference
+  article: https://r-spatialecology.github.io/shar/articles
+
