### 2021-12-01 bbeb25d
[skip-ci] Try smaller font pkgdown
