+++ _pkgdown.yml
@@ -3,6 +3,8 @@ url: https://r-spatialecology.github.io/shar/
 template:
   bootstrap: 5
   bootswatch: minty
+  bslib:
+    font_scale = 12px
 
 reference:
   - contents:
