+++ README.md
@@ -10,6 +10,8 @@
 
 <!-- badges: end -->
 
+<a href='https://r-spatialecology.github.io/shar/'><img src='man/figures/logo.png' align="right" width="150" /></a>
+
 # shar
 
 **S**pecies-**H**abitat **A**ssociations in **R** is a `R` package to
