### 2021-12-02 323e4a3
Misc fixes to JOSS paper
