### 2021-12-02 #3811 @Bisaloo
- I fixed a couple of typos in the manuscript in https://github.com/r-spatialecology/shar/pull/91
- It's a good idea to update your GitHub Actions from upstream from time to time. For example, you could have benefit from the new `clean = TRUE` argument in `pkgdown::deploy_to_branch()`, which would probably have helped a little with your pkgdown problems yesterday.
- If you choose to list the dependencies in the Installation section of your README, you should probably add Rcpp which is a non-base dependency as well
