### 2021-12-02 e490292
:pencil: Some misc improvements to README, homepage, paper
