+++ README.Rmd
@@ -22,7 +22,6 @@ knitr::opts_chunk$set(
 
 <a href='https://r-spatialecology.github.io/shar/'><img src='man/figures/logo.png' align="right" width="150" /></a>
 
-
 # shar
 
 **S**pecies-**H**abitat **A**ssociations in **R** is a `R` package to analyze species-habitat associations. Therefore, information about the location of the species is needed (as a point pattern) and about the environmental conditions (as a raster map). In order to analyse the data for significant habitat associations either the location data or the environmental data is randomized n-times. Then, counts within the habitats are compared between the randomized data and the observed data. Positive or negative associations are present if the observed counts is higher or lower than the randomized counts (using quantile thresholds). Methods are mainly described in Plotkin et al. (2000), Harms et al. (2001) and Wiegand & Moloney (2014). **shar** is mainly based on the [`spatstat`](http://spatstat.org) (Baddeley et al. 2015) and [`raster`](https://rspatial.org/raster/) (Hijmans 2017) package.
@@ -112,6 +111,14 @@ results_habitat_association(pattern = species_a, raster = torus_trans, significa
 
 results_habitat_association(pattern = reconstruction, raster = landscape_classified, significance_level = significance_level)
 ```
+
+## Citation
+
+The `shar` package is part of our academic work. To cite the package or acknowledge its use in publications, please cite the paper.
+
+> Hesselbarth, M.H.K., (2021). shar: A R package to analyze species-habitat associations using point pattern analysis. Journal of Open Source Software, 6(67), 3811. https://doi.org/10.21105/joss.03811
+
+The get a BibTex entry, please use `citation("shar")`.
 
 ## Contributing and Code of Conduct
 
