+++ vignettes/articles/publication_record.Rmd
@@ -8,7 +8,14 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-We try to maintain a list of publications that have used the **shar** package. If you have used **shar** in your own work, we would love to hear from you. Please file an issue on GitHub so we can add your work to this list.
+We try to maintain a list of publications that have used the **shar** package. If you have used **shar** in your own work, we would love to hear from you. Please [open an issue on GitHub](https://github.com/r-spatialecology/shar/issues) or [drop us an e-mail](mailto:mhk.hesselbarth@gmail.com) so we can add your work to this list.
 
-## List of publications
+The `shar` package is part of our academic work. To cite the package or acknowledge its use in publications, please cite the paper.
 
+> Hesselbarth, M.H.K., (2021). shar: A R package to analyze species-habitat associations using point pattern analysis. Journal of Open Source Software, 6(67), 3811. https://doi.org/10.21105/joss.03811
+
+The get a BibTex entry, please use `citation("shar")`.
+
+## A list of publications
+
+...Work-in-Progress...
