### 2021-12-02 #3811 @mhesselbarth
https://github.com/r-spatialecology/shar/commit/e49029242d54503381105fc2594eb95ee511f60c

- I accepted the PR
- I added the `clean = TRUE` argument, but I think thats also the default (?)
- I removed `Rcpp` from the `DESCRIPTION` since `shar` no longer depends on it.
