### 2021-12-02 #3811 @mhesselbarth
All GitHub actions currently fail because https://packagemanager.rstudio.com/ is down (https://github.com/r-lib/actions/issues/447).
