### 2021-12-02 #3811 @Bisaloo
What about the reference written as plain text? Yamada et al. 2006
