+++ paper.bib
@@ -239,6 +239,17 @@
   isbn = {978-1-4200-8254-8}
 }
 
+@article{Yamada2006,
+  title = {Habitat associations of Sterculiaceae trees in a Bornean rain forest plot},
+  author = {Yamada, T. and Tomita, A. and Itoh, A. and Yamakura, T. and Ohkubo, T. and Kanzaki, M. and Tan, S. and Ashton, P.S.},
+  year = {2006},
+  journal = {Journal of Vegetation Science},
+  volume = {17},
+  number = {5},
+  pages = {559--566},
+  doi = {10.1111/j.1654-1103.2006.tb02479.x}
+}
+
 @article{Yang2016,
   title = {Detangling the effects of environmental filtering and dispersal limitation on aggregated distributions of tree and shrub species: Life stage matters},
   author = {Yang, Q.-S. and Shen, G.-C. and Liu, H.-M. and Wang, Z.-H. and Ma, Z.-P. and Fang, X.-F. and Zhang, J. and Wang, X.-H.},
