### 2021-12-02 #3811 @mhesselbarth
https://github.com/r-spatialecology/shar/commit/5ba1f84ead10cd15856826700acc665ee4980799
Ups...thanks for pointing out! That completly slipped my mind.
