### 2021-12-02 #3811 @Bisaloo
Can you please have one last look at the paper for any last remaining typos?

If everything looks good, please tag a new version on GitHub and I can accept your submission :slightly_smiling_face:
