### 2021-12-02 #3811 @mhesselbarth
Fixed some last typos in https://github.com/r-spatialecology/shar/commit/bc93f1a21bb87a8d2472830d9a5fc910ea05a629

In the footer my name is abbreviated by "Hesselbarth, M.H.", however, this should be "Hesselbarth, M.H.K."! Is it possible to fix that somehow?
