### 2021-12-02 #3811 @Bisaloo
Could you try to add a space between `H.` and `K.` [in the yaml header](https://github.com/r-spatialecology/shar/blob/bc93f1a21bb87a8d2472830d9a5fc910ea05a629/paper.md?plain=1#L12) please to make it clear to LaTeX that they are two different words?

I'll dig more and ask around if it doesn't work.
