### 2021-12-02 #3811 @Bisaloo
If it doesn't work, for reference, initials are defined here:

https://github.com/openjournals/whedon/blob/92346a0773a4003bf0ef361b661dc2787f492936/lib/whedon/author.rb#L46-L48
