+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   publication_record: publication_record.html
   reconstruct_multiple_patterns: reconstruct_multiple_patterns.html
   get_started: get_started.html
-last_built: 2021-12-02T16:05Z
+last_built: 2021-12-02T16:12Z
 urls:
   reference: https://r-spatialecology.github.io/shar/reference
   article: https://r-spatialecology.github.io/shar/articles
