### 2021-12-02 #3811 @mhesselbarth
Did not work, but also not really important I think.

I created a tag called JOSS (https://github.com/r-spatialecology/shar/releases/tag/JOSS)
