### 2021-12-02 #3811 @Bisaloo
@openjournals/dev, it looks like whedon is not able to handle very well authors with multiple middle names. Is there anything that can be done?
