+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: 2.7.3
-pkgdown: 2.0.0
+pkgdown: 2.0.1
 pkgdown_sha: ~
 articles:
   background: background.html
@@ -7,7 +7,7 @@ articles:
   publication_record: publication_record.html
   reconstruct_multiple_patterns: reconstruct_multiple_patterns.html
   get_started: get_started.html
-last_built: 2021-12-02T16:12Z
+last_built: 2021-12-03T12:49Z
 urls:
   reference: https://r-spatialecology.github.io/shar/reference
   article: https://r-spatialecology.github.io/shar/articles
