### 2021-12-05 #3811 @arfon
> @openjournals/dev, it looks like whedon is not able to handle very well authors with multiple middle names. Is there anything that can be done?

This will have to be fixed post publishing sorry. Please tag me again here when the paper is published and I'll fix it up.
