### 2021-12-09 #3811 @mhesselbarth
@arfon

Reminder that in the footer my name is abbreviated by "Hesselbarth, M. H. [...]", however, it should be "Hesselbarth, M. H. K. [...]".
