### 2021-12-13 #3811 @arfon
> Reminder that in the footer my name is abbreviated by "Hesselbarth, M. H. [...]", however, it should be "Hesselbarth, M. H. K. [...]".

That should be updated now. Please note the PDF on the JOSS site can take up to 24 hours to update due to caching.
